Extract spokes.Exec
