+++ spokes-receive-pack.go
@@ -5,10 +5,7 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"os/signal"
-	"syscall"
 
-	"github.com/github/spokes-receive-pack/internal/governor"
 	"github.com/github/spokes-receive-pack/internal/receivepack"
 	"github.com/github/spokes-receive-pack/internal/spokes"
 )
@@ -36,24 +33,5 @@ func mainImpl(stdin io.Reader, stdout, stderr io.Writer, args []string) (int, er
 		return 0, nil
 	}
 
-	ctx, stop := signal.NotifyContext(ctx, syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP)
-	defer stop()
-
-	rp, err := spokes.NewSpokesReceivePack(stdin, stdout, stderr, args, BuildVersion)
-	if err != nil {
-		return 1, err
-	}
-
-	g, err := governor.Start(ctx, rp.RepoPath)
-	if err != nil {
-		return 75, err
-	}
-	defer g.Finish(ctx)
-
-	if err := rp.Execute(ctx, g); err != nil {
-		g.SetError(1, err.Error())
-		return 1, fmt.Errorf("unexpected error running spokes receive pack: %w", err)
-	}
-
-	return 0, nil
+	return spokes.Exec(ctx, stdin, stdout, stderr, args, BuildVersion)
 }
